requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2
django-appconf==1.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fire==0.5.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
rcssmin==1.1.2
requests==2.31.0
rjsmin==1.2.2
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
